// src/app/main/apps/procesos/ascenso/models/ascenso.model.ts
export class CalificacionDetalleModel {
    numeroDocumento: number;
    abreviaturaDocumento: string;
    descripcionGrupoCompetencia: string;
    apellidoPaterno: string;
    apellidoMaterno: string;
    nombres: string;
    codigoPlaza: string;
    puntajeFinal: number;
    escalaObtenida: string;
    descripcionRegimen: string;
    descripcionGrupoModalidad: string;
    abreviaturaRegimenLaboral: string;
    numeroDocumentoCompleto: string;
    codigoModular: string;
    centroTrabajo: string;
    cargo: string;
    escalaMagisterial: string;
    areaCurricular: string;
    jornadaLaboral: string;
    descripcionInstancia: string;
    descripcionSubinstancia: string;
    descripcionRegimenLaboral: string;
    descripcionPlaza: string;
    puntajeEvaluacion: number;
    puntajeTrayectoria: number;
    puntajeBonificacion: number;
    puntajePUN: number;
    descripcionRegion: string;
}